refresh layout

+++ application/controllers/admin/Role.php
@@ -65,9 +65,10 @@ class Role extends BaseController
 
     // breadcrumbs
     $items["breadcrumbs"] = $this->_breadcrumbs();
+
     // render view html
-    $output["content"] = $this->load->view($this->view_list, $items, true);
-    $this->load->view("layouts/Main", $output);
+    $output = $this->load->view($this->view_list, $items, true);
+    $this->response($output);
   }
 
 
@@ -101,10 +102,9 @@ class Role extends BaseController
     // select box HERE !!
 		$items["selectBoxStatusId"] = $this->StatusModel->selectBox("statusId", @$statusId);
 
-
     // render view html
-    $output["content"] = $this->load->view($this->view_form, $items, true);
-    $this->load->view("layouts/Main", $output);
+    $output = $this->load->view($this->view_form, $items, true);
+    $this->response($output);
   }
 
 
@@ -185,6 +185,12 @@ class Role extends BaseController
 
     if (!$action || !$id) show_404();
     if ($action != "__delete__") show_404();
+
+    // verify user active this role
+    if ($this->BaseModel->isRemoveable($id) == false){
+      redirect($this->route . "?____CANNOT_DELETE_USER_EXISTS____");
+    }
+
 
     $results = $this->BaseModel->delete($id);
     if ($results->status)
